script/requirements.txt
python-docx==0.8.10
selenium==3.141.0 